Have Pyboard read from env var BELAY_DEVICE if device not specified
