+++ belay/helpers.py
@@ -2,11 +2,6 @@ import importlib.resources
 import secrets
 import string
 from functools import lru_cache, partial, wraps
-from typing import List
-
-from serial.tools import list_ports
-
-from belay.usb_specifier import UsbSpecifier
 
 from . import snippets
 
@@ -28,26 +23,3 @@ def random_python_identifier(n=16):
 def read_snippet(name):
     resource = f"{name}.py"
     return importlib.resources.files(snippets).joinpath(resource).read_text()
-
-
-def list_devices() -> List[UsbSpecifier]:
-    """Lists available device ports.
-
-    Returns
-    -------
-    List[UsbSpecifier]
-        Available devices identifiers.
-    """
-    devices = [
-        UsbSpecifier(
-            vid=port.vid,
-            pid=port.pid,
-            serial_number=port.serial_number,
-            manufacturer=port.manufacturer,
-            product=port.product,
-            location=port.location,
-            device=port.device,
-        )
-        for port in list_ports.comports()
-    ]
-    return [x for x in devices if x.populated()]
