+++ belay/usb_specifier.py
@@ -1,4 +1,4 @@
-from typing import Dict, Optional
+from typing import Dict, List, Optional
 
 from pydantic import BaseModel, Field
 from serial.tools.list_ports import comports
@@ -48,7 +48,7 @@ class UsbSpecifier(BaseModel):
         spec = self.dict(exclude_none=True)
         possible_matches = []
 
-        for port_info in comports():
+        for port_info in list_devices():
             if _dict_is_subset(spec, vars(port_info)):
                 possible_matches.append(port_info)
         if not possible_matches:
@@ -62,4 +62,29 @@ class UsbSpecifier(BaseModel):
         return possible_matches[0].device
 
     def populated(self):
+        # some ports, like wlan and bluetooth on macos,
+        # don't populate any meaningful fields.
         return bool(self.dict(exclude_none=True))
+
+
+def list_devices() -> List[UsbSpecifier]:
+    """Lists available device ports.
+
+    Returns
+    -------
+    List[UsbSpecifier]
+        Available devices identifiers.
+    """
+    devices = [
+        UsbSpecifier(
+            vid=port.vid,
+            pid=port.pid,
+            serial_number=port.serial_number,
+            manufacturer=port.manufacturer,
+            product=port.product,
+            location=port.location,
+            device=port.device,
+        )
+        for port in comports()
+    ]
+    return [x for x in devices if x.populated()]
